scripts/requirements.txt
google-generativeai
google-genai
python-dotenv
pillow
opencv-python
numpy

